move encoded features to prototype datasets namespace
