+++ torchvision/prototype/datasets/utils/__init__.py
@@ -1,3 +1,4 @@
 from . import _internal  # usort: skip
 from ._dataset import Dataset
+from ._encoded import EncodedData, EncodedImage
 from ._resource import GDriveResource, HttpResource, KaggleDownloadResource, ManualDownloadResource, OnlineResource
